Success CRUD

+++ mvc/app/Views/admin/products/list.php
@@ -1,3 +1,6 @@
+<div class="message">
+    <?= $message ?>
+</div>
 <table border="1">
     <tr>
         <th>#ID</th>
@@ -19,7 +22,7 @@
             </td>
             <td>
                 <a href="<?= ROOT_PATH ?>product/edit?id=<?= $pro->id ?>">Edit</a>
-                <a href="<?= ROOT_PATH ?>product/delete?id=<?= $pro->id ?>">Delete</a>
+                <a onclick="return confirm('Bạn có muốn xóa không?')" href="<?= ROOT_PATH ?>product/delete?id=<?= $pro->id ?>">Delete</a>
             </td>
         </tr>
     <?php endforeach ?>
